release: v4.1.0 - third round deep audit (86 bugs)
Legacy fixes (11): execve argv docs, capability architecture notes,
vfs inode refcount docs, signal blocked check verified, alloc_pages
identity mapping docs, MB1/MB2 detection docs, pipe_close ops table
comparison, fat32 LFN TOCTOU comment, explain bounds checks

New CRITICAL (1):
- NC1: map_page sets PTE_USER on intermediate page table entries

New HIGH (23):
- pagetable: COW free on refcount zero, atomic PTE overwrite
- sched: non-BSP CPU idle loop checks rq->count
- signal: sigreturn RFLAGS masking (clear IOPL/NT/TF/AC)
- pipe: set blocked_reader before releasing lock
- tss.S: TSS RSP0 initialized to stack_top
- keyboard_handler.S: IRQ1 EOI to slave PIC (0xA0)
- nvme/virtio_blk: uint64_t multiplication overflow guards
- net: total_len cap, dns ancount cap, dns name bounds
- syscall: sys_mmap kernel space overlap check
- capability: permission check docs
- ext2: inode_size > block_size check, dir lookup/readdir OOB
- journal: circular buffer tail fix

New MEDIUM (28): atomic ref operations, lazy allocation VMA validation,
vruntime actual consumption, yield vruntime update, add_child lock,
TLB shootdown comment, SMAP leaf PTE only, pipe_close priv first,
smp IPI mfence, nvme NULL checks, virtio_net MAC size check,
squashfs buffer free/block_list check, fat32 uint64 dir size,
tcp_cong cwnd overflow, seccomp/capability limitation docs

New LOW (23): check_resched atomic, NX preservation, kmalloc obj_size
zeroing, dhcp retransmit docs, ipv6 extension docs, tcp_cong RTT=0 skip,
sprintf bounds/%u/%x/%i, efi fallback docs, mod_hello version format,
loongarch64 boot.S andi fix, greet_count unsigned, module dep_names loop,
shared lib ASLR docs, nanosleep EINTR docs

Version: v4.1.0

+++ kernel/fat32.c
@@ -721,10 +721,10 @@ static int fat32_find_free_slot(struct fat32_sb_info *sbi,
 /*
  * Compute the total size of a directory by walking the cluster chain.
  */
-static uint32_t fat32_get_dir_size(struct fat32_sb_info *sbi,
+static uint64_t fat32_get_dir_size(struct fat32_sb_info *sbi,
                                    uint32_t dir_cluster) {
     uint32_t cluster_size = sbi->cluster_size;
-    uint32_t total_size = 0;
+    uint64_t total_size = 0;
     uint32_t cluster = dir_cluster;
 
     while (cluster >= 2 && cluster < FAT32_CLUSTER_EOC_MIN) {
@@ -1754,8 +1754,8 @@ struct super_block *fat32_mount(struct block_device *bdev) {
         }
 
         /* Compute root directory size */
-        uint32_t root_size = fat32_get_dir_size(sbi, sbi->root_cluster);
-        root_inode->size = root_size;
+        uint64_t root_size = fat32_get_dir_size(sbi, sbi->root_cluster);
+        root_inode->size = (size_t)root_size;
         ((struct fat32_inode_info *)root_inode->priv)->file_size = root_size;
 
         /* Create the super_block */
